address of operator can now behave as an lvalue

+++ include/ast/lists/declaratorList.hpp
@@ -13,7 +13,7 @@ public:
     using NodeList::NodeList;
     void generatePython(std::ostream &os, PythonContext &context, int scopeDepth = 0) const override;
     void generateIL(std::vector<Instr> &instrs, ILContext &context, std::string destReg) const override;
-    int getDeclCount() const;
+    size_t getDeclCount() const;
     DeclaratorPtr getDeclarator(int index) const;
 
 protected:
